# lib/api-zod/src/generated/types/eventsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PumpFun Dev Fee Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { WalletEvent } from "./walletEvent";

export interface EventsResponse {
  events: WalletEvent[];
  total: number;
}