# lib/api-zod/src/generated/types/trackedWallet.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PumpFun Dev Fee Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface TrackedWallet {
  address: string | null;
  label: string | null;
  isTracking: boolean;
  lastChecked: string | null;
}