# lib/api-zod/src/generated/types/walletEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PumpFun Dev Fee Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { WalletEventType } from "./walletEventType";

export interface WalletEvent {
  id: number;
  type: WalletEventType;
  txHash: string;
  amount: number;
  amountUsd: number | null;
  description: string;
  category: string | null;
  timestamp: string;
  walletAddress: string;
}