# lib/api-zod/src/generated/types/walletStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PumpFun Dev Fee Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface WalletStats {
  totalFeesClaimed: number;
  totalDexBoosts: number;
  totalBoostSpend: number;
  totalTransfers: number;
  lastActivity: string | null;
  walletBalance: number | null;
}